chore: remove `--docs` option from `cloud-init schema` (#5857)
It has been broken and has marginal value, so it is not worth
maintaining. Since it was a developer-facing option, this should
have no backwards compatibility concerns.

Fixes GH-5756
